requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
python-dotenv>=1.0.0
supabase>=1.2.0
python-multipart>=0.0.6
jinja2>=3.1.2
requests>=2.31.0
itsdangerous>=2.1.2
starlette>=0.27.0
pydantic>=2.4.2 